+++ src/OrderBy/OrderByDirective.php
@@ -6,7 +6,6 @@ use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\InputValueDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
 use Illuminate\Support\Str;
-use Nuwave\Lighthouse\Schema\AST\Codegen;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
 use Nuwave\Lighthouse\Schema\AST\PartialParser;
 use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
@@ -14,9 +13,12 @@ use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirectiveForArray;
 use Nuwave\Lighthouse\Support\Contracts\ArgManipulator;
 use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
+use Nuwave\Lighthouse\Support\Traits\GeneratesColumnsEnum;
 
 class OrderByDirective extends BaseDirective implements ArgBuilderDirective, ArgDirectiveForArray, ArgManipulator, DefinedDirective
 {
+    use GeneratesColumnsEnum;
+
     public static function definition(): string
     {
         return /** @lang GraphQL */ <<<'SDL'
@@ -27,8 +29,16 @@ directive @orderBy(
     """
     Restrict the allowed column names to a well-defined list.
     This improves introspection capabilities and security.
+    Mutually exclusive with the `columnsEnum` argument.
     """
     columns: [String!]
+
+    """
+    Use an existing enumeration type to restrict the allowed columns to a predefined list.
+    This allowes you to re-use the same enum for multiple fields.
+    Mutually exclusive with the `columns` argument.
+    """
+    columnsEnum: String
 ) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
 SDL;
     }
@@ -70,11 +80,10 @@ SDL;
         FieldDefinitionNode &$parentField,
         ObjectTypeDefinitionNode &$parentType
     ): void {
-        if ($allowedColumns = $this->directiveArgValue('columns')) {
+        if ($this->hasAllowedColumns()) {
             $restrictedOrderByName = $this->restrictedOrderByName($argDefinition, $parentField);
             $argDefinition->type = PartialParser::listType("[$restrictedOrderByName!]");
-
-            $allowedColumnsEnumName = Codegen::allowedColumnsEnumName($argDefinition, $parentField);
+            $allowedColumnsEnumName = $this->generateColumnsEnum($documentAST, $argDefinition, $parentField);
 
             $documentAST
                 ->setTypeDefinition(
@@ -83,9 +92,6 @@ SDL;
                         "Order by clause for the `{$argDefinition->name->value}` argument on the query `{$parentField->name->value}`.",
                         $allowedColumnsEnumName
                     )
-                )
-                ->setTypeDefinition(
-                    Codegen::createAllowedColumnsEnum($argDefinition, $parentField, $allowedColumns, $allowedColumnsEnumName)
                 );
         } else {
             $argDefinition->type = PartialParser::listType('['.OrderByServiceProvider::DEFAULT_ORDER_BY_CLAUSE.'!]');