+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Avoid erasing the model information from the wrapping paginated results `type` when defining
   a paginated `@hasMany` field after a field with `@paginate` https://github.com/nuwave/lighthouse/pull/1149
 
+### Added
+
+- Add optional `columnsEnum` argument to the `@whereConditions`, `@whereHasConditions` and `@orderBy` directives https://github.com/nuwave/lighthouse/pull/1150
+
 ## [4.8.0](https://github.com/nuwave/lighthouse/compare/v4.8.0...4.7.2)
 
 ### Added